# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Rosadisk Agent API
  description: API service for Rosadisk Agent
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /v1/disks:
    get:
      summary: List available disks
      operationId: ListDisks
      tags:
        - disks
      responses:
        "200":
          description: List of disks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Disk"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/fs:
    get:
      summary: List btrfs filesystems
      operationId: ListFilesystems
      tags:
        - filesystems
      responses:
        "200":
          description: List of btrfs filesystems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Filesystem"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a new btrfs filesystem
      operationId: CreateFilesystem
      tags:
        - filesystems
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFilesystemRequest"
      responses:
        "201":
          description: Filesystem created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Filesystem"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/mounts:
    get:
      summary: List mounted btrfs filesystems
      operationId: ListMounts
      tags:
        - mounts
      responses:
        "200":
          description: List of mounted btrfs filesystems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MountInfo"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Mount a btrfs filesystem by UUID
      operationId: MountFilesystem
      tags:
        - mounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MountRequest"
      responses:
        "201":
          description: Filesystem mounted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MountInfo"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/subvolumes:
    get:
      summary: List subvolumes
      operationId: ListSubvolumes
      tags:
        - subvolumes
      responses:
        "200":
          description: List of subvolumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subvolume"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a new subvolume
      operationId: CreateSubvolume
      tags:
        - subvolumes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubvolumeRequest"
      responses:
        "201":
          description: Subvolume created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subvolume"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/subvolumes/{id}:
    get:
      summary: Get subvolume details
      operationId: GetSubvolume
      tags:
        - subvolumes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Subvolume details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subvolume"
        "404":
          description: Subvolume not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a subvolume
      operationId: DeleteSubvolume
      tags:
        - subvolumes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Subvolume deleted successfully
        "404":
          description: Subvolume not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/subvolumes/{id}/snapshots:
    get:
      summary: List snapshots for a subvolume
      operationId: ListSubvolumeSnapshots
      tags:
        - snapshots
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SnapshotInfo"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/subvolumes/{id}/backups:
    get:
      summary: List backups for a subvolume
      operationId: ListSubvolumeBackups
      tags:
        - backups
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of backups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BackupInfo"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/config:
    get:
      summary: Get global configuration
      operationId: GetConfig
      tags:
        - config
      responses:
        "200":
          description: Global configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlobalConfig"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Update global configuration
      operationId: UpdateConfig
      tags:
        - config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GlobalConfig"
      responses:
        "200":
          description: Updated global configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlobalConfig"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/job_logs:
    get:
      summary: List job logs
      operationId: ListJobLogs
      tags:
        - jobs
      parameters:
        - name: job_type
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of job logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobLogSummary"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/job_logs/{id}:
    get:
      summary: Get a job log with output and error
      operationId: GetJobLog
      tags:
        - jobs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job log details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobLogDetail"
        "404":
          description: Job log not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    Disk:
      type: object
      required:
        - name
        - size
        - type
      properties:
        name:
          type: string
          description: Device name (e.g., sda, nvme0n1)
          example: sda
        size:
          type: integer
          description: Size in bytes
          example: 500107862016
        type:
          type: string
          description: Device type (disk, loop)
          example: disk
        vendor:
          type: string
          nullable: true
          description: Device vendor
          example: ATA
        model:
          type: string
          nullable: true
          description: Device model
          example: WDC WD10EZEX-08L
        fstype:
          type: string
          nullable: true
          description: Filesystem type (e.g., ext4, btrfs)
          example: btrfs
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: failed to list disks
    Filesystem:
      type: object
      required:
        - uuid
        - size
        - devices
        - raid_profile
      properties:
        uuid:
          type: string
          description: Filesystem UUID
          example: 550e8400-e29b-41d4-a716-446655440000
        label:
          type: string
          nullable: true
          description: Filesystem label
          example: data-pool
        size:
          type: integer
          description: Total size in bytes
          example: 1000000000000
        devices:
          type: array
          description: List of device paths
          items:
            type: string
          example: ["/dev/sda1", "/dev/sdb1"]
        raid_profile:
          type: string
          description: RAID profile (single, raid0, raid1)
          example: raid1
    CreateFilesystemRequest:
      type: object
      required:
        - devices
        - label
        - raid_profile
      properties:
        devices:
          type: array
          description: List of device paths to create filesystem on
          minItems: 1
          items:
            type: string
          example: ["/dev/sda1", "/dev/sdb1"]
        label:
          type: string
          description: Filesystem label
          pattern: '^[a-zA-Z0-9][a-zA-Z0-9-]*$'
          maxLength: 255
          example: data-pool
        raid_profile:
          type: string
          description: RAID profile (single, raid0, raid1)
          enum: ["single", "raid0", "raid1"]
          example: raid1
    MountInfo:
      type: object
      required:
        - uuid
        - label
        - mountpoint
        - devices
        - used
      properties:
        uuid:
          type: string
          description: Filesystem UUID
          example: 550e8400-e29b-41d4-a716-446655440000
        label:
          type: string
          description: Filesystem label
          example: data-pool
        mountpoint:
          type: string
          description: Mount point path
          example: /mnt/rosadisk/550e8400-e29b-41d4-a716-446655440000
        devices:
          type: array
          description: List of device paths
          items:
            type: string
          example: ["/dev/sda1", "/dev/sdb1"]
        used:
          type: integer
          description: Used space in bytes
          example: 500000000000
    MountRequest:
      type: object
      required:
        - uuid
      properties:
        uuid:
          type: string
          description: Filesystem UUID to mount
          pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
          example: 550e8400-e29b-41d4-a716-446655440000
    DefragConfig:
      type: object
      required:
        - enabled
      properties:
        enabled:
          type: boolean
        frequency:
          type: string
          enum: ["daily", "weekly", "monthly"]
    Subvolume:
      type: object
      required:
        - id
        - name
        - fs_uuid
        - path
        - created_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        fs_uuid:
          type: string
          format: uuid
        path:
          type: string
        compression:
          type: boolean
        quota:
          $ref: "#/components/schemas/QuotaConfig"
        snapshots:
          $ref: "#/components/schemas/SnapshotConfig"
        backups:
          $ref: "#/components/schemas/BackupConfig"
        defrag:
          $ref: "#/components/schemas/DefragConfig"
        nfs:
          type: boolean
        smb:
          type: boolean
        created_at:
          type: string
          format: date-time
    QuotaConfig:
      type: object
      required:
        - enabled
      properties:
        enabled:
          type: boolean
        limit:
          type: integer
          description: Required when enabled is true
    SnapshotConfig:
      type: object
      required:
        - enabled
      properties:
        enabled:
          type: boolean
        frequency:
          type: string
          enum: ["hourly", "daily", "weekly", "monthly"]
        retention:
          type: integer
          description: Number of snapshots to keep
    BackupConfig:
      type: object
      required:
        - incremental
        - full
      properties:
        incremental:
          $ref: "#/components/schemas/BackupSchedule"
        full:
          $ref: "#/components/schemas/BackupSchedule"
...
    CreateSubvolumeRequest:
      type: object
      required:
        - name
        - fs_uuid
        - compression
        - defrag
        - nfs
        - smb
        - quota
        - snapshots
        - backups
      properties:
        name:
          type: string
          pattern: '^[a-zA-Z0-9][a-zA-Z0-9_-]*$'
          maxLength: 255
        fs_uuid:
          type: string
          format: uuid
          description: Parent filesystem UUID from GET /v1/fs
        compression:
          type: boolean
        quota:
          $ref: "#/components/schemas/QuotaConfig"
        snapshots:
          $ref: "#/components/schemas/SnapshotConfig"
        backups:
          $ref: "#/components/schemas/BackupConfig"
        defrag:
          $ref: "#/components/schemas/DefragConfig"
        nfs:
          type: boolean
        smb:
          type: boolean
    GlobalConfig:
      type: object
      required:
        - backup
        - snapshot
        - defrag
        - scrub
        - balance
        - backup_storage
        - encryption
      properties:
        backup:
          $ref: "#/components/schemas/VolumeJobSchedule"
        snapshot:
          $ref: "#/components/schemas/VolumeJobSchedule"
        defrag:
          $ref: "#/components/schemas/VolumeJobSchedule"
        scrub:
          $ref: "#/components/schemas/DiskJobSchedule"
        balance:
          $ref: "#/components/schemas/DiskJobSchedule"
        backup_storage:
          $ref: "#/components/schemas/BackupStorageConfig"
        encryption:
          $ref: "#/components/schemas/EncryptionConfig"
    VolumeJobSchedule:
      type: object
      required:
        - enabled
        - time
      properties:
        enabled:
          type: boolean
        time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          example: "03:00"
          description: 24-hour format HH:MM
        hourly_minute:
          type: integer
          minimum: 0
          maximum: 59
          example: 0
          description: Minute of the hour for hourly frequency
        weekly_day:
          type: string
          enum: ["monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday"]
          example: monday
          description: Day of week for weekly frequency
        monthly_day:
          type: integer
          minimum: 1
          maximum: 31
          example: 1
          description: Day of month for monthly frequency
...
    JobLogSummary:
      type: object
      required:
        - id
        - job_type
        - status
        - started_at
      properties:
        id:
          type: integer
        job_type:
          type: string
        mountpoint:
          type: string
          nullable: true
        subvolume_id:
          type: string
          nullable: true
        target_name:
          type: string
          nullable: true
        status:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    JobLogDetail:
       allOf:
         - $ref: "#/components/schemas/JobLogSummary"
         - type: object
           properties:
             output:
               type: string
             error:
               type: string
    SnapshotInfo:
      type: object
      required:
        - id
        - subvolume_id
        - name
        - path
        - frequency
        - size
        - created_at
      properties:
        id:
          type: string
          format: uuid
        subvolume_id:
          type: string
          format: uuid
        name:
          type: string
        path:
          type: string
        frequency:
          type: string
        size:
          type: integer
        created_at:
          type: string
          format: date-time
    BackupInfo:
      type: object
      required:
        - id
        - subvolume_id
        - type
        - snapshot_path
        - path
        - size
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        subvolume_id:
          type: string
          format: uuid
        type:
          type: string
        parent_id:
          type: string
          format: uuid
          nullable: true
        snapshot_path:
          type: string
        path:
          type: string
        size:
          type: integer
        upload_details:
          type: string
          nullable: true
        status:
          type: string
        error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true